Move encodings to a utility folder
