+++ src/commands/utility.js
@@ -2,7 +2,7 @@ import hashjs from 'hash.js'
 import { base64 } from 'rfc4648'
 
 import { command, UsageError } from '../command.js'
-import { base58, utf8 } from '../encoding.js'
+import { base58, utf8 } from '../util/encoding.js'
 
 function hmacSha256(data, key) {
   const hmac = hashjs.hmac(hashjs.sha256, key)
