+++ src/util/encoding.js
